Added timestamps in Chatlog
It was a little odd that timestamps checkbox in Preferences affected
only regular chat messages, but not the chatlog, so I decided to add
timestamps into the chatlog.

I hate to use `const_cast` to add a timestamp to a received message, so
if someone knows a better solution, don't hesitate to mention it.


Former-commit-id: b790eaacb6e583d230a932c5b9a011fc098bb6da
Former-commit-id: 3467c5a3d49accca29eab26652e2bd32f4434d2d
Former-commit-id: 0f6887d85daa83df6bdc653b1adb150d349fe199
Former-commit-id: 968a1e7d3f2755e40fc9068d3e1ea033b41c4f2d [formerly 7f19e375efceec8f51616ebcab1c4b219510f065]
Former-commit-id: 815f8dd853bf9d59926980fa8c26466bfbc7e13f