Refactoring unit class to use boost::scoped_ptr instead of raw pointer.
This change aims to simplify the code, making the intents more explicit.
There wasn't a memory leak in the previous version, but it could easily be
introduced in the future.

Also, RAII is part of the coding standards of Wesnoth, as noted in the wiki.

There are other raw pointers in the unit class, but unit don't own them, then
I leave them unchanged.


Former-commit-id: f9657adca303e0558cb9fcf231be837d58001a1a
Former-commit-id: 4b86c708dafb3781700eee23a3d19ddfaa14ca7f
Former-commit-id: 1b9d3dfaa1adf5cdbd911e91381ebffcf20ec143
Former-commit-id: e6043d7353b38e73734d6c6452460b1a2fb19fef [formerly 6434cb1a95e003ae5e36aed2a3c1212eb772dae5]
Former-commit-id: 2298e9fe990bb81a01c3f567c5014b091f65fba3