Prevent unchecked memory access in cut_surface()
The result of create_compatible_surface() may be encapsulating a null
pointer whenever the underlying call to SDL_CreateRGBSurface() fails --
in particular, if the width or height specified are negative and not
equal to -1, which is handled specially.

Without this check, using the ~CROP(x,y,w,h) image function with
negative width or height not equal to -1 resulted in a SIGSEGV as soon
as the create_compatible_surface() result was used.


Former-commit-id: eadc574d84b8d0bcbbd3443ba85bdfa0d95ff463
Former-commit-id: 72a8a357baf311a2763f667c2f728a125a02c967
Former-commit-id: 57c423638e980d7156f02c1256a57c221b4ff7b0
Former-commit-id: 74fad0b5140bba01f38181ade5bd99228b861386 [formerly a80360a338e7c6aaa7b8b396be8621d4549993cd]
Former-commit-id: 6fabdd045b1338870200cfaa5de3d8a3295cebec