Prevent unchecked memory access in cut_surface()
The result of create_compatible_surface() may be encapsulating a null
pointer whenever the underlying call to SDL_CreateRGBSurface() fails --
in particular, if the width or height specified are negative and not
equal to -1, which is handled specially.

Without this check, using the ~CROP(x,y,w,h) image function with
negative width or height not equal to -1 resulted in a SIGSEGV as soon
as the create_compatible_surface() result was used.


Former-commit-id: d7ee953b8273908627b1ffab4045b85683ce486b
Former-commit-id: 5c4c24fefb6950f8ece65b8ee34479c57e08cf76
Former-commit-id: a29786b5b00dd230852e696be37685274ee91301
Former-commit-id: c7baa15c8a35abbb038a0573e7d83cd6f0ab19b7 [formerly b02e7d1e061c7d6f5f0d86c66cae959ccfad17a6]
Former-commit-id: fca29efcc0caae383241333602062fa5411d3099