campaignd: Allow existing add-ons with invalid names to be reuploaded
This change is for backwards compatibility purposes only in order to
allow authors to continue uploading new versions of add-ons with
previously valid names that have been rendered invalid by the whitelist
addon_name_legal() method. This commit should be reverted next time an
add-ons server instance is deployed (hopefully for 1.12 first instead of
1.13.x).

A warning is provided in the logs for admins whenever an add-on with an
invalid name is reuploaded.


Former-commit-id: d9c94fba0157e8c9b027ff7828f7b524362cb4b7
Former-commit-id: 84e6964cb4cff5a77b99775ea4e28746f3dc5d94
Former-commit-id: 8a18b08773cacc98480f753d712bacda8eb5cada
Former-commit-id: 9ed1823c18514ee80c0472b05c036d622c925284 [formerly d4ad9abe74607fa7ac5060da2540d28019263ff9]
Former-commit-id: 499e10a61de654c8479916ee64a52fa341460c07