campaignd: Allow existing add-ons with invalid names to be reuploaded
This change is for backwards compatibility purposes only in order to
allow authors to continue uploading new versions of add-ons with
previously valid names that have been rendered invalid by the whitelist
addon_name_legal() method. This commit should be reverted next time an
add-ons server instance is deployed (hopefully for 1.12 first instead of
1.13.x).

A warning is provided in the logs for admins whenever an add-on with an
invalid name is reuploaded.


Former-commit-id: 8fccce5c2379ba8a3c115f5df472a6b8da975de4
Former-commit-id: 1c9678dbfc0c2006027121810f7c13b73eeb6f5d
Former-commit-id: 1d487b549b993736f7df6f25a58cd2b10e306c93
Former-commit-id: 9d3afb1223fdceee839884bae4db59a638409f6d [formerly edf1c5c717353a0620483ca49c6d7dc07cda82cf]
Former-commit-id: 4f089adf3a64e0a4edcb45cf0bf25f92037ad196