addon/info: Demote circular dependency errors to info loglevel
It turns out this is actually very useful according to a certain add-on
author who is definitely not me.

For example: have Campaign Foo and Music Pack Foo. Music Pack Foo can
depend upon Campaign Foo so a player attempting to download it is
suggested to download Campaign Foo as well since it's entirely useless
otherwise. Meanwhile, Campaign Foo depends upon Music Pack Foo in order
to be able to use some custom music in the campaign proper.


Former-commit-id: e2ed5a295549ab129fee26a7a4a73f002aaedc4b
Former-commit-id: 7b087fe02ffabbbac7c382ea203d326d21a167c1
Former-commit-id: cedd3711c21dfca72185ed9895036f4592dd1d57
Former-commit-id: c826022d090f13ee3f57c0be47ea1aec01694bfe [formerly 9e9ae90864065da10f1ab358ab0dc139fa9fe961]
Former-commit-id: ed55f744daf3b27d3c3850ba256540d203b25d54