give line number of DEFENSE_ANIM too, when warning that get_hit_sound is also present
Building on our earlier fix pointing this message's line number to the get_hit_sound, it would be even more useful to have the line number of the DEFENSE_ANIM as well. To get this, we change the has_defense_anim variable's value from False to None, and True to i. It is possible that a unit might have two macros for male and female variants, in which case we will assume that the get_hit_sound is most likely associated with the first.


Former-commit-id: 18ecc18d030861cd2c3a71f8bde1ec179f111b95