give line number of DEFENSE_ANIM too, when warning that get_hit_sound is also present
Building on our earlier fix pointing this message's line number to the get_hit_sound, it would be even more useful to have the line number of the DEFENSE_ANIM as well. To get this, we change the has_defense_anim variable's value from False to None, and True to i. It is possible that a unit might have two macros for male and female variants, in which case we will assume that the get_hit_sound is most likely associated with the first.


Former-commit-id: c03eb446a21d67348caba977efb6d93fde1a4976
Former-commit-id: 3b44ac9762b8cd2a0b47f581cae9377bbb74cc2c
Former-commit-id: 7072a7548ceedf02d5c0ab1b0b7099d33354cbae
Former-commit-id: bcb2d6a284ce72711fbd17a39a01b38467144bdd [formerly 31c316a8bcbb85ab33f91586fcdcc07806ab93f7]
Former-commit-id: 1fb3f8f11e60ad1fbda8966c9e7e5f15deb5c26e