clean up get_hit_sound messages
The first message has a couple of problems. Technically, get_hit_sound is not a tag, and there is a stray quote mark at the end. Also, i+1 points to the line number of the [/unit] tag, which is not particularly helpful information. This can be changed to point to the line of the get_hit_sound attribute.

For the second message, the %d get_hit_sound is an index position, so +1 for the line number.


Former-commit-id: f12fa9255463da7dc6e4d27c81a3e4c2eea1cf98
Former-commit-id: 9a26c2ae85006e4d6a803788c7a2cf02b6fa5aaa
Former-commit-id: f9d669f0e6de5c69e51a4b77dfdc2eb3e6ff015c
Former-commit-id: a9eb25fb09d8aa0c3ab18fa2c611ab99e6ebe719 [formerly e5ae6b70f18f207cb978fd1dadbea79a6ddd7492]
Former-commit-id: 82ffe29e06a389cbaa485f3740434cde3924981b