Add combo selection for campaign's entry points to mp::configure.
The selection is only available when there is more than one scenario
which belongs to campaign and has 'allow_new_Game' set to 'yes'.

The combo items are named using 'new_game_title' attribute from
[scenario] tag. If attribute does not exists, 'name' attribute is used
as a fallback.

This commit is for gui changes only. The entry point selection itself
is not yet functional.


Former-commit-id: eeb2fd865ab3deb217bac0c1fb463a2de168b9c8
Former-commit-id: d05b847d3f8b0e3da8de3884a9d9ffc4d3d6e4e6
Former-commit-id: 2778c17c43d525c7243c1f286fa54d6566de9049
Former-commit-id: 18c1ba6f826cdd6d14c4839c21ca4a24f08666cb [formerly 304fbea66779c0364061a0beb06d90a15174b7aa]
Former-commit-id: b70371ca405157ff369e32dfc07725c6863549a7