New [campaign] tag attributes to specify the number of players.
The attributes are 'min_players' and 'max_players'. Their default value
is '2'. If 'max_players' is greater than 'min_players' - 'max_players'
would be equal to 'min_players'.

The number of players for campaigns is displayed similarly as for
stand-alone scenarios, except that if 'min_players' is not equal
'max_players', the number is displayed in 'min to max' format.


Former-commit-id: 1bbf014d2fc277e8f3d72537d7fa7b24d9ad581e
Former-commit-id: 0ed834fcfe4170158a7c32996f74b2b744f65b63
Former-commit-id: bea5223cd3b8201e32be20a48b2baa0a8e3ef33a
Former-commit-id: 0f366a842aabe5fe34b836664b96c8c51fa4264e [formerly dc3b0c978cb4743cd6e5a75ae41769cac521224b]
Former-commit-id: 90cb7854e4b39187319da8a390362fe6a2b7065b