incorporate old image_defensive into DEFENSE_ANIM creation and discard, or warn of its presence
While getting rid of the deprecated get_hit_sound, we can get rid of the even more hoary image_defensive.

We do this with new variables that are initially set to None. While in_unit, we look for the 'image_defensive=' key. If encountered, we record the line's index position (image_defensive) and its value (defend_image).

When we hit [/unit], we look back to see if there is already a DEFENSE_ANIM. If so, and its reaction image matches the value recorded in defend_image, we figure there is no need to preserve the old key, and enter it into a list of image_defensive attributes to be deleted (image_done). If we are creating a DEFENSE_ANIM, we use defend_image for the reaction image instead of doubling up the base image. Once this is done, again there is no point in keeping image_defensive around, and it is entered into image_done for deletion.

If neither of these cases is met, we offer warnings that an outdated key is in use.

When all of the file's lines have been iterated through, we can then remove those image_defensive lines that have been marked as unnecessary.


Former-commit-id: 21f6217d7691ab344428e5faa25a58da863cb5f0
Former-commit-id: f0ac8435306bca9b70cf8890e16283edd766f096
Former-commit-id: 1b7d5a3506b84461871ce3adaec37128050e06b4
Former-commit-id: dbb541d481fadbc0b814e508652ff414513616bc [formerly 136ad7e177403804e33f4ba61de2b620223df9d3]
Former-commit-id: b1aa6e77660abf2f4db7bfcac03965342609264f