check for presence of [defend] animation with DEFENSE_ANIM
We look for a [frame] inside a [defend] tag. If there is no image_defensive, we also use any image as our defend_image. We warn if there is already a DEFENSE_ANIM. If there is a get_hit_sound, we go through with the transformation, but alert the porter.


Former-commit-id: 381ea549c8e2623cbda71e5e303d2b1a939fca43