check for presence of [defend] animation with DEFENSE_ANIM
We look for a [frame] inside a [defend] tag. If there is no image_defensive, we also use any image as our defend_image. We warn if there is already a DEFENSE_ANIM. If there is a get_hit_sound, we go through with the transformation, but alert the porter.


Former-commit-id: 793efa4ebe3033f6eef1f87f6e154f015dd5f3dd
Former-commit-id: c805365898f3419f7ccda90423e26e18a1202a02
Former-commit-id: 3f82caf683127fc80d617a2a99c6c810d7f4bf01
Former-commit-id: 5ff929cee141a46577e820e4d10f4ba45b7d9050 [formerly 8f09f9441a27e15b00098e619d26fa431f8f8e62]
Former-commit-id: 58b6e7da689239b80cb2561279922e7bb313d97e