bugfix newline problems in output
I had noticed that the line replacing the get_hit_sound with DEFENSE_ANIM didn't have a newline, but assumed that it was part of 'comment'. Nope!

There are also two lines where a misplaced quotation mark led to an extraneous space being added to the end of a line.


Former-commit-id: c31e36165af93cecb8803ee02f2767b6cb0f6040
Former-commit-id: 1d78e618976ad847971d5fac3488ba5f6f82ed39
Former-commit-id: 42ded26687a0c39ce7af11dad0c4174d9e40494a
Former-commit-id: dc045c0a2397d74eaa604996db03ef170b2feb52 [formerly fb4833c90178dda637f29f863e4c17b5008e3525]
Former-commit-id: b6ff75e05b28ec2fe42bbd7f8acfb8361556e301