bugfix newline problems in output
I had noticed that the line replacing the get_hit_sound with DEFENSE_ANIM didn't have a newline, but assumed that it was part of 'comment'. Nope!

There are also two lines where a misplaced quotation mark led to an extraneous space being added to the end of a line.


Former-commit-id: 8418d10668a678ba56fefd7e0ea73b690ea001bc