bugfix insertion of attack's description= key
First, the newline is added to "description = " rather than "new_line = ". But description was only changed if it didn't begin with a quotemark, meaning that those that *did* start with a quote weren't getting a newline.

Second, new_line was supposed to inherit indentation through "leader(syntactic)", but the line had already been stripped before "syntactic", in "fields = ".


Former-commit-id: 9838558ba1a2f40131616b1041ce07b2f5951302
Former-commit-id: 64906bd0baeeaeead5a00e40707af587f9306303
Former-commit-id: a13b64d2cb0192e2f87e3b4d8996d5e4dffd1bd4
Former-commit-id: b93dd1a311c513f0cb433feb6ff374290b3ab58a [formerly d39782e2ed507d0e3cd41a29a708dac22bc502b1]
Former-commit-id: 469f5cc51c9c92c6a5d749047a8347b1923c7edb