move upconversion of ancient abilities ahead of lifting [frame] to fix problem of weapon specials appearing after [frame] sequence
After my last change, I noticed a puzzling failure by wmllint to convert a weapon special. This special was among some attributes that followed the [frame] sequence. It seems that Python does not wait for the earlier code block to complete before running the new one, and those lines aren't passed through the new block because they've been deleted and stashed in 'postframe'. When they're spewed back out, the new block has already passed those lines by.

I was relieved to find that this was not an issue introduced by my change, but an existing one. When I ran the original wmllint on the file, I found that the special= line got deleted, without being replaced by the [special] tags and macro. The latter is supposed to appear when wmllint hits the [/attack] tag, but never triggers because [/attack] has been changed to [/attack_anim].

Moving this code block up, so that abilities and specials are transformed before the [frame] lift (and 'postframe' stash), appeared to fix the problem. Hopefully, it won't cause a new on to show up.


Former-commit-id: d5537082f7b738c08c8a33a5526caa9bcd244708
Former-commit-id: 0386671c1213b78643847b8618313d7d9c2ff092
Former-commit-id: 6340866137804a5cc63806a4b578aaa89ced4114
Former-commit-id: f6d57cb428adda94b6233f64b8bbe5c1b8588064 [formerly e7cdf0b756224bd415fc17873577e22c0d1bce84]
Former-commit-id: 14c5715b76781e266e53d21fd2de23576a09577d