Refactor faction set up in multiplayer.
The code has been improved to switch from index lookups to lookups by
ids. This makes the implementation more flexible for future improvements
and easier for maintenance.

The "Custom" faction has been added. It is automatically chosen and
locked for sides with a recruit list. The sides without recruit list
won't see it in their faction choice combo box.

The recruit list in side will be now be overriden by a faction if map
settings are not in use.


Former-commit-id: c01a083eb15db9ce0cc558d7876cbaf8f4c88bc0
Former-commit-id: a807a76c2b408d405c13cb1a2a8ce8e89a8c39c1
Former-commit-id: ce6057e6ae7589bc0d26d090130b97a354d92111
Former-commit-id: 023ea7bab854429abe46c2a9ba400ace545a6e82 [formerly 980b55019965cbaadcb362e285cd8ab944f24157]
Former-commit-id: f3190beab3f0a53c23721bd4face362bf8c2ee32