Added a check to change a unit's ellipse depending on its ZoC
This modification allows the game to automatically select the unit's ellipse, by checking if said unit emits a ZoC; if the unit loses or gains it, due for example to a level-up or to the Stun ability, the ellipse will change accordingly. This effectively makes no longer useful declaring the ellipse in a unit's .cfg file.


Former-commit-id: 1776ad6e3228c915df65741309f141fd7357f608
Former-commit-id: 7ae5ec6e7fd17657257d947c49a0cb7383de14a9
Former-commit-id: 956f8b4ed3d6ff826fa6dd1bdae4bf97de1537d2
Former-commit-id: 7dcb1f891bed6ad9c7d15f19ea9bba3a7a36bc98 [formerly ed46a59639272f25d3c8f5351a0677cf0eae7d7f]
Former-commit-id: a01581a3874317d51ea3f51ece8387bc9ac06322