Renamed hero and leader nozoc ellipses to make them compatible with the automatic ellipse management
I had to rename these ellipses to make them compatible with the IS_HERO macro as well. This way, a user can assign a ellipse with the usual ellipse=misc/ellipse-hero, and when a unit loses its ZoC, it'll be updated accordingly. Leaving the file names as they were wasn't a good solution, because it required a way to split the ellipse-hero string to insert the nozoc- part.
With this solution, no WML change is required; however, we may want to add a wmllint rule to warn UMC authors if they have a ellipse= key in their unit cfgs.


Former-commit-id: 6bf4837a0b4b0cae791cc3bee074bac1bf18e9b5
Former-commit-id: b5557de9eaca16da655fe70f6d3516930a580a99
Former-commit-id: 3bfce2bde40f01322d9ce65708d6db7b651fa964
Former-commit-id: cccf85e5a31fd17cd1f88c59d48bcf2b9e54869c [formerly 3c61b8558bea4d3e77db9d2e29e286aabfad61cf]
Former-commit-id: 528c6a8f8bf43cf046c2ddbaa73836b7399f4ef5