Removed no longer useful ellipse= key from mainline units (bats)
Since ellipses automatically change depending on the ZoC emission, there is no point to change them manually in the cfgs - we don't want the engine to look for a misc/ellipse-nozoc-nozoc file. This commit is the first of a series that will fix all mainline units, even the campaign-specific ones.


Former-commit-id: 019ed4b57f5b1b9d06ae1ed6dae7f01de54878a3
Former-commit-id: 064aa68d9fa120a3db9ff8780b1816aa2e3ae395
Former-commit-id: dc1a3aa402863840714cc22c5a073308517cba65
Former-commit-id: 80f705bbabb929117e6496894c45a7b77d8f0bd3 [formerly 2f5df221d8d5f9ff971e79319540004d6db0298e]
Former-commit-id: b13790960e99dca09e5b0c014ac0e4b0a6ce64cf