preserve sound_miss value with SOUND:HIT_AND_MISS macro
Some [sound] tags include a sound_miss key as well. To preserve this data, I created two new variables to capture the sound_miss= and time= values, and use them to fill out the SOUND:HIT_AND_MISS macro. Because the macro is inserted before [frame] instead of after, 'if soundpath' needs to be moved up before the insertion so that the macro ends up under [attack_anim], rather than in the middle of [attack].


Former-commit-id: 363d6efff52542a0865ab80be2a6cc38f14e7d12
Former-commit-id: 4a4e06bed4f2d3dd17387cf50454f081dbfcd316
Former-commit-id: 94408ce25ed0875ea102313fec07b56c6bda5dc1
Former-commit-id: d1ccd1ebcaf09a40a0ed0369c741e7dbb5b8bf97 [formerly dfdda255eecdbf0fb1efd086b72098f079219856]
Former-commit-id: 776caea69dc110f7ec9d96b0a060578f56243bc3