ugly workaround for another MSVC compile error
The problem is that MSVC requires n in char buf[n] to be a constant; it
cannot be computed previously. It needs to be given as a number directly
(char buf[10];) or be a const var (const int n =10; char buf[n];)
It's ugly because of manual new/delete but other than that I'd have to
dive deeper into the code it seems. If the code is being worked on, please
remove the manual memory management and, for instance,
use std::string or stringstream for the get_map() callstack instead.


Former-commit-id: 3857cc23d8ad72668b5c843d69ecbd338540a031
Former-commit-id: b86fb35cf180efafa1016678ac0c590ff2b0ff05
Former-commit-id: 34dfb0ffe8bc0b6688f6ab5d085b730433a460e3
Former-commit-id: f2c8e7d761306b95b1c6fb480527b1349b880b9b [formerly 964cd9b517a9ac100b164c082d29873158d7a9df]
Former-commit-id: 001dd65bd8a5eaaa455fe4be9a14f2d1526a13d9