try to match sound's time key with a [frame] begin time
Rather than automatically putting the soundpath in the first frame, we collect a list of begin= keys and their locations. When it's time to convert, we check to see if any of those begin times match with the sound time, and if so, put sound in that [frame]. Only if there is no match do we default to the first frame, and we log a message noting the lack of a match.


Former-commit-id: 083b5a3f6df1d18fd7fa479d3a79741f570ae3e4
Former-commit-id: 6b5b89c5c5d93fbc893a58159958ea7bf1b6d23a
Former-commit-id: 95f598b64982c063c71e15bf923a1b35bc1e4f1a
Former-commit-id: 3a9cb4ca3e44a6fa90a0731054bebfa8e30462fb [formerly 1e929c391c870eac0ce064060cc848faefc274d9]
Former-commit-id: a0c1c62ac9cdeb99762376cf2486a1d98011a5d8