try to match sound's time key with a [frame] begin time
Rather than automatically putting the soundpath in the first frame, we collect a list of begin= keys and their locations. When it's time to convert, we check to see if any of those begin times match with the sound time, and if so, put sound in that [frame]. Only if there is no match do we default to the first frame, and we log a message noting the lack of a match.


Former-commit-id: 62508ce12b221f480ef2a51a69b4511dec84a528