wmllint: extend Elvish Hunter's Windows sys.argv fix to cover more cases
While testing my next commit, I discovered that EH's fix works when there is only one argument, or if the offender is the last argument, but doesn't work with multiple entries. His fix is meant to work on each argument, but the (unintentionally) escaped quote no longer serves to end the argument, causing following arguments to be considered part of the same argument.

Using split() allows us to break apart these misconjoined arguments. With rstrip(), we prevent an empty string that Windows will also complain it cannot find. However, if there are three or more arguments, there will still be lumped-together arguments unless all arguments up to the second from last also end with a backslash and quote. It is impossible to cover every possible case.

The re.sub handles the probably rare case where a backslash before a quote comes within the argument rather than at the end. However, it will only work if there is only one argument.

All this is unnecessary if the OS is not Windows (also, I haven't had the opportunity to test this on a non-Windows system to see if it has any side-effects there). So I've put it under a sys.platform condition.


Former-commit-id: 369e696a333b4698ea219ec26c28ad0fa015930a
Former-commit-id: 85de25b174c6c62ac5b943199c21d057d396b766
Former-commit-id: b54887e14ef5cf2279ed4339e95c8bc7895b2fc0
Former-commit-id: 90723d793226f950e39b3dc22596c1031057b9c3 [formerly a1f3be01d3b6c5f0e80ba027cb0b50dc759507fd]
Former-commit-id: f219a6bff71fcbb8de87f9ca7cda5cfa95a4b8b8