wmllint: extend Elvish Hunter's Windows sys.argv fix to cover more cases
While testing my next commit, I discovered that EH's fix works when there is only one argument, or if the offender is the last argument, but doesn't work with multiple entries. His fix is meant to work on each argument, but the (unintentionally) escaped quote no longer serves to end the argument, causing following arguments to be considered part of the same argument.

Using split() allows us to break apart these misconjoined arguments. With rstrip(), we prevent an empty string that Windows will also complain it cannot find. However, if there are three or more arguments, there will still be lumped-together arguments unless all arguments up to the second from last also end with a backslash and quote. It is impossible to cover every possible case.

The re.sub handles the probably rare case where a backslash before a quote comes within the argument rather than at the end. However, it will only work if there is only one argument.

All this is unnecessary if the OS is not Windows (also, I haven't had the opportunity to test this on a non-Windows system to see if it has any side-effects there). So I've put it under a sys.platform condition.


Former-commit-id: afe045563d4df9d890b2360d10fec7cbbda62b8f
Former-commit-id: f0287a7c777304500b1f7d30afe1ef058d224d82
Former-commit-id: cd540141569bb127d7a43997a3aca5b5583dc69a
Former-commit-id: 1686fa16757a3e4df6f051f7234efb0b88d517e1 [formerly fc979787f93321dfbc98363705aced4d441e2a75]
Former-commit-id: 861043b99a56df175a00c1a69cbb64fccb4e1670