sys.exit(1) if globbing matches no files/directories
After more testing, I realized that I did not take into account the possibility that the wildcard pattern would not match anything. In that case, the following 'if not arguments' clause would run wmllint on the entire current directory - which could very well be something that you do not want!


Former-commit-id: 2afd8c2960331cfbee936ceaa4880cf4126fb1e3
Former-commit-id: 937a750559bd6e3d78c716f209f8c7a035c19539
Former-commit-id: 82a6f7227f81528fc2fedeb53f8a23e7526b3f01
Former-commit-id: 46e3121e4e787836a1ef3303edd6a5a1cea856f3 [formerly 8974e25af422f301abb89436d6c64ad7217d9138]
Former-commit-id: 7db858822a6dd868bf9a70752d92ca07f455a6f8