wmllint: fix bug with ifdef_stack
This section assumed that "#ifdef" and "#endif" would come at the very start of a line. When an author would indent the #ifdef but not the #endif, ifdef_stack.pop() would kill the starting value of None, leaving an empty list. wmllint would then crash:

  File "wmllint", line 1138, in global_sanity_check
    recruit[ifdef_stack[-1]] = (i+1, map(lambda x: x.strip(), value.split(",")))

IndexError: list index out of range

Stripping the line not only stops the crashes, but allows wmllint to pick up #ifdefs that it wasn't before.


Former-commit-id: 74e6a8023fbca3b894bc8ddbcd14b61f83c4ed65
Former-commit-id: 2387f0b562ced37d711609d3d07d72b13bf42400
Former-commit-id: 739f20d715bf484229eb5a2dfc7ce2e631c211c7
Former-commit-id: 9d68a23f15ce76fe2cace56907a1246a17f8aca9 [formerly b35fb55c0697aaaa21173d4a43855c66e7a41968]
Former-commit-id: c6a3647cb2384a7947587a3679daf65e19e0ae94