Reserve sides for players who played them in previous scenario.
The information for the previous scenario's sides which have "save_id"
and a player ("player_id") assigned to them are now being recorded in
"mp_game_settings". This information is used in mp::connect for every
non-last scenario to reserve sides for previous players so that
they could easily join the game and be placed in the correct side.


Former-commit-id: 35c5b3821802c693db015f16a8690a98bbc64d9c
Former-commit-id: 88a3c3caeef4d64cce51244b8a880404587884b7
Former-commit-id: 248252581bedfc4b0b0617a237d0d3f2a3113d6d
Former-commit-id: 60600e8cbbb6dbf284d1580c516cf4ed65ca5782 [formerly d05db694cbee855fd71ed20c4d44c54599fddd59]
Former-commit-id: 63c9ce06d401a47582e015934360c3e027d82a94