recognize characters from "wmllint: who" macros and append to present
Now that we see how the whopairs are recognized, we can see that the magic comment accepts a comma-separated list, for macros that deal with more than one character.

We also see that if it is necessary to remove a character who leaves the party, this can be done with another entry prefixed by double minuses.


Former-commit-id: f21ccfe92a7b746a2fe15844572d3ec51bba0e6d
Former-commit-id: baf842331ad254a277fb0e3059941cabfe042a8b
Former-commit-id: ea474be0e58269af8e961525ee168bb5c48be1f0
Former-commit-id: 34195370d854aa02fac8371248b4c6b1eaaeba42 [formerly 531d3f3603b0e69f0ecdd9ec83cc814b61a5436f]
Former-commit-id: a08390c2026115fa046a4d73c80d69a2cbd2433e