wmllint: add switch -K --known for suppresing consistency_check()
I used "inconsistency" for the actual variable name, because "known" seems more likely to be accidentally reused.

I pondered whether to allow the scenario check to go forward, but decided to just make a clean break.

Note that this does not prevent any of the information-gathering for the consistency check, just the check itself.

Why would you want to use this option? Of course, you should run the consistency check at some point. But if you simply want to recheck if you've fixed all the bugs in your campaign, you might not want to have wmllint slog through data/core again.


Former-commit-id: 23152468ebfe315047c7d4db1e10a2a1968ab2a5
Former-commit-id: 68dfd60f60bc593300905311f15d9cb9c041cb88
Former-commit-id: 8bcc9fe4182233527bd546c0f6d1d91b6e81eb5d
Former-commit-id: 9e2ef74cfe541b77b9dec7191b4d0cc1c168fe75 [formerly 55fd663065e6ff1a93f08b72b4ed420888af2397]
Former-commit-id: 54af64d769fe974808f74459650887e7f40acb65