Have pump() no longer pump events raised while it is running.
The main way I know of for getting into this situation is hitting
the recursion maximum within pump(), and if that happens you probably
have runaway WML. So it should be better to move on instead of
continuing to process events.
(Normally, events raised within an event handler are pumped before
that event handler exits, so would not be affected by this change.)

I think the old behavior is an artifact of the old implementation,
but I'm making this a separate commit in case removing this causes
problems (easier to diagnose, especially if I'm not the one diagnosing).


Former-commit-id: 809099630a49ca082c63b51eae67c7fe1a8f6daf
Former-commit-id: 2477b4cca663c0d4b5c16473025f75d3bd2bbf5b
Former-commit-id: 7ca0797477cd7d6bd381f94b64c46e25284e016f
Former-commit-id: 2f42b48303b533cad1c17dcf2f115892ed86eaec [formerly ce56666c02c912370032058b526cc770b59c610a]
Former-commit-id: 49db51016119221642d5abc1e6cd2939096b1d5f