wmlindent: fix pango indentation-stopping bug (pointed out by vultraz)
I had a hunch that the hash would be the problem, because I realized that wmllint usually just looked for '#' to figure out if a line contained a comment, and that this also matched pango color coding. My wmllint backslash/userdata commits used this same re.split match to prevent this false positive.

(It won't prevent all unintended matches, e.g. "Guard #4", but I don't think that can be helped.)

The problem came, I believe, when the old line interacted with the block ten lines down:

        elif eligible.count('"') % 2:
            dostrip = not dostrip

The value's closing quote was on the other side of the split.

Former-commit-id: 765f36c70039633ecfea2f7e427447194a43297a
Former-commit-id: 434155f6c885468134962fce3382ecb86bdb817e
Former-commit-id: d37884c1967b7a17bb368b6508169730c4353e7b
Former-commit-id: 5deebc2b029b6287ad0b59e1d5928087089d696f [formerly f1f4e082d6257d1dc16a158db46dbd2456c793e7]
Former-commit-id: dba3b64d3c086bccd3e76b5993884ec968d9deff