wmlindent: fix pango indentation-stopping bug (pointed out by vultraz)
I had a hunch that the hash would be the problem, because I realized that wmllint usually just looked for '#' to figure out if a line contained a comment, and that this also matched pango color coding. My wmllint backslash/userdata commits used this same re.split match to prevent this false positive.

(It won't prevent all unintended matches, e.g. "Guard #4", but I don't think that can be helped.)

The problem came, I believe, when the old line interacted with the block ten lines down:

        elif eligible.count('"') % 2:
            dostrip = not dostrip

The value's closing quote was on the other side of the split.


Former-commit-id: a733974319724244ae2dcb1efec7a76d90e4dfd4
Former-commit-id: a808bfd89202e3bcf1a1b2f47b04747cba9baf83
Former-commit-id: c105ac41e40c4b9ba342efc596caf7169a0f7d53
Former-commit-id: 508976a524e126f3553d9313519277d865f31686 [formerly b84fa5d7e76658afa87a35562b8be4b5a09141f5]
Former-commit-id: 0a682cf20db227aab7c47096a0afcc23c950a118