fix wmltools' parse_attribute to not consider pango color hashes as a comment
While there are many places that wmllint assumes that "#" begins a comment, from Vultraz's description it sounded like parse_attribute was being used to find the attribute's value, and then a string_strip was done on the value. So I looked for the direct cause in wmltools.

Again, the solution is to look for whitespace to precede the hashsign.

Incidentally, I don't know why the original code had the first "where -= 1", then had value and comment go from [:where+1] and [where+1:].


Former-commit-id: 4b7c5a86edc6d6dcbf173d1be32890fb77c364bd
Former-commit-id: b3c4774c4018c0b5c920fd7f84c0fb2c9855a19d
Former-commit-id: 4b39eb2fdf400666db9626d69aa957a0efe324cc
Former-commit-id: d3793793e5a6f299cc45b92547387f63c8f61ffe [formerly f90c7ef771784d0c4ea380a1440231c3cf3a7a0e]
Former-commit-id: 744061a8691c70bea5e27552d72972fa5b48fd4b