check for textdomain paths without '/translations', and binary paths referring to 'external' data
The convention that [textdomain] uses "/translations" is strongly established, and I can't think of a legitimate reason for an add-on not to be following it.

The binary path check is a crude test. The names that took hold for menu image directories are "/public" and variations of "/external*", so we look for those strings. It does not catch the worst case of all - when all binaries are outside the campaign define, not just a set-aside directory.

I first thought of these checks while brainstorming ways to use the in_textdomain and in_binary_path code in hack_syntax(). However, realizing that these checks did not really hack any syntax, I wanted to find someplace in the sanity checks where the code would fit. I finally found it.

Former-commit-id: 83260f588914982690471d99204250be4995767d
Former-commit-id: 6f9744cfd6eb1f53a89dffa9b82b9b83c081db44
Former-commit-id: 9dd7c4aebe7e093d5299e1559d2f1d6b609d165f
Former-commit-id: a86853e49953e3f96b342a974dab04775f2a7cda [formerly 84bd4292f3950c5bcd0f515b00d22f89f384cfb6]
Former-commit-id: e6f98e44226923b6f42fc2415f1d89d9c5df8901