addon/mg: Trap config::error exceptions thrown during .pbl read
This introduces a new minimal exception object to encapsulate some
relevant information (the file path and the config::error message) so it
can be propagated to the manager UI code correctly.

This is the first step is ensuring faulty .pbl files are not reported
during connection to the add-ons server as a "network communication
error".


Former-commit-id: 0435c6f395c8cb6ba8a0c2ae3b96e6cf1cc65f54
Former-commit-id: 5ab0fd5333a2088bf6c36e40dfbdacd4fc237c4e
Former-commit-id: b86a1559c498647ea8f1aebc234a3fbd79249179
Former-commit-id: e947fec19a7ff6f7df117be846598964f8f3a40c [formerly 962234143599d11685da14219c7b03c7ac163097]
Former-commit-id: bc955ec26000af94d957d2f7a765ed57ec22da73