Speed up AI turns 5-10% by deferring tod creation
Profiling --no-gui runs (I used a commercial tool) showed a surprising
amount of time is spent inside terrain_filter::match_internal. This
appears to mostly be as a result of checking every single terrain tile
to determine where the leader location is, which seems ripe for
optimization itself.

However, within this function 5-10% of the time is spent creating a
time_of_day that is rarely used by these AI matches. Simply reordering
the lines to defer its creation until inside the if block that checks
whether tod is a factor eliminates this part of the overhead.

This patch speeds AI turns up by 5-10% on my test machine (an 8-core
Linode 2048).


Former-commit-id: 35eae131f5bf36ff3c37373ae59c6d62c5949a09
Former-commit-id: 72b586fa2f3185e7220b8050e8d46901d4733f7a
Former-commit-id: 4ccc5901bef7469976696bec9cdb8d5f2dcc8244
Former-commit-id: a19d9c0dbe15feaeab2fe9ac81f40cc4c98b8852 [formerly be6a60cd37099fdffcb92de5490b34f55fc0a185]
Former-commit-id: c2372dec5a2f2a9785feb4db7cfd95449a30f2e5