Add the function utils::might_contain_variables().
This function does a quick, but imprecise, check to see if a string
might change after variable interpolation. If this function returns
false, it is safe to assume the string will not change.

I thought I saw another function that does this, but I cannot find it.

The usecase I have in mind is where a string will persist for a while,
perhaps for an entire scenario, where optimizations could be made if
the string was constant, and where variables would be uncommon (with
"uncommon" relative to the size of the optimizations). For example,
event names seem to fit.


Former-commit-id: c7194d840dd899f0d09f593a084994541e0cedec
Former-commit-id: fbb4c0a1d207bbad580ad4e3000a6393691d2dc3
Former-commit-id: d8a196c1f9a433ac3f6463a151693430ea0231e2
Former-commit-id: a081fb8ad7c6a2012f5931fab13da244b02b9359 [formerly 65d7cc220bc92270f1ef27ec5a169554c41f02ec]
Former-commit-id: e00189e80be739be6fe0ff4e037ffaa84fc82c63