Use "Use map settings" to lock options in mp::connect.
A new "force_use_map_settings" attribute has been added to disallow
playing with custom settings in order to avoid broken gameplay
experience. The attribute's default value is "no", unless mp::connect is
is opened after winning a scenario. In that case, the attribute
determines whether "Use map settings" are in use or not.


Former-commit-id: 8943095f8c7630f2ef91788ab96e8636d9ee7ef5
Former-commit-id: 40f197bb59e41888a1a01fa622184b4c85cc44c5
Former-commit-id: 3028b1c747ba7768361d1847b0a5770c24175a75
Former-commit-id: b851984aac8ddf643f1cbd2f87eaa28e81358a0e [formerly 0a028fa66fda1cc718b79120d9db3ffeb2f51fbe]
Former-commit-id: 80c8a4aadfb3532777fe536192b828a5c9dc1d35