Always use mp::connect_engine to set up an MP scenario.
The engine was slightly modified to have an option to force import
players when starting a game, so that previous players would be carried
over and correctly assigned to their corresponding sides.

As a side effect, the play_game() loop received a slight clean up,
because some code there became unnecessary.


Former-commit-id: 99a3ba8ac627e83c51e75116a7710130500e0d50
Former-commit-id: c3fa4afd32ddd400e5cd5350c8b3696c81d0161c
Former-commit-id: 1c12ab53e62c032a8a6dab1175e19bc70c367b37
Former-commit-id: 0aadee1c89163eb470ce62a00c0602924385793b [formerly 2543d3173e3145e925c1543b90b56063e35417af]
Former-commit-id: d00278c72b4e2fa10b33439383a0de52e2876b1e