Make it possible to overwrite a default leader by mp::connect selection.
Default leader is found by checking for "id" in [side] tag. If id was
found, it will check if leader was carried over and is now in [unit]
tag. If leader was still not found, it will try to pick a first unit
which can recruit.

If "Use map settings" are off or no leader was specified, mp::connect
will allow to select a leader from a faction. If default leader was in
carryover and the new leader was selected, the old leader attributes
will be cleared, keeping only an "id" and a newly set "type" and
"gender" attributes.


Former-commit-id: e44cebe6a589f6051d283f4eea8f5027d7298821
Former-commit-id: 1b8f68befc6ff93df3c4c2719f30969e509af4c9
Former-commit-id: ac65d2c6a06e17e20f919969c2619b0dd230b852
Former-commit-id: bb0fb58bbe8b84474c8f5e0e6ccda4c6135701c2 [formerly e643ddf4a7b02525395aaf50227fd632710b691d]
Former-commit-id: 606f5a76b92062e0d55ea7748210301b3ccbab75