note when recruitment mismatches involve non-standard usage classes
When recruitment patterns include bogus usage classes, it will trigger a message during the consistency check. However, this check won't tell you whether your problem is with the usage classes or the recruits. For example, if you get a message that "no light fighter units are recruitable", your first thought might be that you didn't include a light fighter unit in the recruit list. The message doesn't clue you in directly that the light fighter simply doesn't exist in mainline.

Now a clause will be added to the message when non-mainline recruitment classes are involved, alerting designers when bogus usage classes are involved, but mild enough that it hopefully won't scare away authors legitimately seeking to use custom classes.

This replaces the old usage check. The next commit will add a magic comment for appending custom usage types.


Former-commit-id: d349551f957c97ddc83cd981322e11b892c95ea2
Former-commit-id: 8a4df43fa9056ee8373c9ed6a6da8ed0f18ce2fe
Former-commit-id: 8a1e8997e85380eb92c6a40263d55af4de460f30
Former-commit-id: 6d88cfbe6bdf04a61685e489b98acf82f774ed5e [formerly d82238c204d63a74365d77b7f03e6bb1a7867fac]
Former-commit-id: c47cc6df8f2e545c0dd93b32588bc31c32f6055d