gui2/tgame_paths: New dialog for displaying game paths in Preferences
The button to invoke this dialog is currently in Preferences -> General.
This will change as soon as I think of a better place/someone else
suggests a better place for the button.

The goal is to allow users to quickly inspect the various paths used by
the game to locate resources and configuration, for debugging or content
creation purposes. The dialog provides options to copy directory paths
to the system clipboard and browse them using the platform's file
manager software, when available.


Former-commit-id: a26f8e06a499b4aefeceaa9f77a33b9e803b3b31
Former-commit-id: d13755e37cf9ce57b69f54f9c1182c4e4eccaab0
Former-commit-id: 0b337374d9714fd707d4fa03d082a358a8c9ddd9
Former-commit-id: 88d1f8fe2c8f49abe69ae5a3fd9989ef90dfa618 [formerly 3a08ec2759f9dc27ddb004957788519f3c1fbf7e]
Former-commit-id: 6760e9ab181476d329c2499d5807b28d725f4279