Add functionality to launch the platform's file manager for a path
This supports X11 platforms, Windows, and OS X. The code was
developed and tested on Debian GNU/Linux, Windows XP, and Windows 7.

The X11 and OS X versions of the function are identical and use an
exec/fork/wait pipeline with the only platform-dependant variable
being the external launcher binary (xdg-open on X11, open on OS X). I
opted for this somewhat ugly approach instead of system() because of
the many dangers resulting from allowing the system shell to
reinterpret the command line at whim.

The Win32 version uses ShellExecute() to launch the default
application associated with the Folder object type, in case Windows
Explorer is not the default handler for whatever reason.

The OS X code path was preliminary tested by mattsc and that's really
all I know about it. More testing on that platform would be very,
*very* welcome, although it is expected to behave pretty much the
same as Linux/X11


Former-commit-id: 9c722f8987fb7bc998827d09bd91daf2c7724a19
Former-commit-id: 6a33d32b70d84b3b3a50f97ad0cd8315f958d18b
Former-commit-id: 96306b5ccc2a6ccc223a1e1eede8197a03e71452
Former-commit-id: a540e355d9bfaac0b91fa18ea256e4088dbbdc4d [formerly 3a24df2accad75f5d43231ae56afa7b6f66e0582]
Former-commit-id: 76f59ff88b0700f1ded22615e2a701d1a05f70a0