Updated "mp_game_settings" after MP scenario is played.
Now mp::lobby is able to display a correct scenario name, number of
turns and map_data for all scenarios in a campaign.

Some "mp_game_settings" are still not being updated to their correct
values after an MP scenario. For this to happen, mp::configure could be
splitted to gui and engine classes and the new engine class could be
used along with mp::connect_engine to update those values.


Former-commit-id: 6eb9d7f088b206fd73ada12393df648b349f971b
Former-commit-id: 71f0444552074eeedf38de9ff57f7e4a26b7a150
Former-commit-id: 39120b1354c9efa06d961841be97ecf385c83c31
Former-commit-id: 9cc39e6e67b8727f89abe67418c70532e5a89205 [formerly 6ad2564b82f1fe3c3ca9b8d802ca8a6765be2bda]
Former-commit-id: 4535292d4fa71b8a15ddb7bf721d12fba7b0ead5