Improve i18n of system tray notifications on Windows
The system tray icon tooltip and some notification headers were not
translatable, which resulted in them being hardcoded to English. Fixed
this by making them translatable.

Caveat: the game_display::send_notification method is designed around
comparing some of the relevant strings. This means that some
functionality may not work properly if the translation language is
changed between notifications. It doesn't seem like anyone considered
this when designing the platform-neutral interface of this
functionality, though, seeing as how there's a call to
game_display::send_notification elsewhere that already uses a localized
argument.

Former-commit-id: 5b4db0dfc847e29843916adad9e7eeb4cca15e9b
Former-commit-id: 49c75d613b4b587806a3e44632c03f0bf46f2192
Former-commit-id: 9796b04d6a9fe99065444a15eb3a425da936ad54
Former-commit-id: 630df9bfcd02e086f3359ef698a6d0faf4b223e0 [formerly 2c7044053b44fd6d31eec055e512d3aff8604dfb]
Former-commit-id: b665b7771f81d31d635f205a53b786f7e7caa1be