Improve i18n of app notifications on OS X
Some relevant user-facing strings were not marked as translatable.

Caveat: the game_display::send_notification method is designed around
comparing some of the relevant strings. This means that some
functionality may not work properly if the translation language is
changed between notifications. It doesn't seem like anyone considered
this when designing the platform-neutral interface of this
functionality, though, seeing as how there's a call to
game_display::send_notification elsewhere that already uses a localized
argument.

Former-commit-id: 6811d30e0e571dbbb37b7af91b31836d66c33eda
Former-commit-id: fabcc7dc881aafda878866f9ad7483e4f9b69f43
Former-commit-id: def1814cd9b2f3247ecc0750af6aec5611838f61
Former-commit-id: a2d268ed3c946aeaa46bc107b5ec066b52fe13df [formerly 6c0aaa138f8cc437b0f38a628a255e90292e4f9d]
Former-commit-id: 0dc8a4ec4d771e085d177cb0a8fc27db8022b61b