Fixed "current_player" being assigned incorrectly (mp::connect_engine).
1st case:
Overwriting "current_player" with default value produces an incorrect
side config and causes the server to not recognize the player if its
username is not the same as "current_player" value. This is not
desirable, since a different player might be assigned to an ex-reserverd
side or a player might just simply change his/her username.

2nd case:
Using default value for "current_player" if no player has taken that
side, results in setting wrong controller in case side was reserved for
host.


Former-commit-id: 432109bdfdf35400503e5c3ee6c9dbf3c5270b68
Former-commit-id: b2bff5a08566ef8e22523fcfc90418d437121d5a
Former-commit-id: 84fdf8e370dac59b521fcf7fd730b4d27c1475f4
Former-commit-id: e20466735797276d79ba9f52368fbf07935030b6 [formerly 034514b61cb664049f0dfe0b70530ae15c865e1f]
Former-commit-id: f7202f518ed097322648152780f30b2326d14eca