addon/mg, gui2/taddon_uninstall_list: Implement reading of add-on titles from _info.cfg
This makes it so the user-facing parts of the UI display the actual
add-on titles from the add-ons server that were stored in the _info.cfg
files when installing. I had to move more code around than I wanted for
this because gui2::taddon_uninstall_list needs to handle not just add-on
ids and names, but also list item number mappings since GUI2 doesn't
allow attaching additional data to listbox items as far as I know.

It would be desirable to allow displaying the actual unedited paths or
ids for reference purposes for the user again (e.g. manual removal of
problematic add-ons) later, without disrupting the standard workflow.


Former-commit-id: 16f7a4703b9dd0f02012f33dae2181f843cdd7ef
Former-commit-id: 4016c52d131d6ff802d69b80e85c47bddd7aa1c8
Former-commit-id: 5f8dcf38a3048219e570eaa0e35d3b3920345d28
Former-commit-id: 3340c7a9444527eb7246b4270bfc4e425e70b6bd [formerly bc2d4bf29e08aef7a8340825a78a16f00a4d60cf]
Former-commit-id: 8a9ab22485d6bc7bc5e55a932918a6bd559cc80f