Reload game config in mp::wait.
The game config is being reloaded to make sure that players have the
same config as host, if possible.

Note that reloading only happens when it is necessary and it would not
happen if host's cache defines matches the player's ones. I.e. regular
MP scenarios won't trigger a config reload.


Former-commit-id: 3c653733342eac64c2ab4b42e02af3a364041988
Former-commit-id: c314d3c96ee0806e4dece944ab627da5a7c80181
Former-commit-id: 8354347d6be80ae6bec3637a804368b5adf3c8a5
Former-commit-id: f2d9baea53c092c2f0e0d96845c079d77f31bf5f [formerly 826e1cdd78df1c6e73ff23b1da1f8eb9c6e0800d]
Former-commit-id: ad4dcbf76b0ad7b31adda78c2a32dbd2691d896a