variation_id replaces the former variation_name.
The new [variation] attribute variation_id is now used to identify the
subtype of the variation. This was done by the variation_name until now.
variation_name is now used to display a translated string to the user.

During creation of the unit tree a new attribute base_ids is stored to
hold a list of the [base_unit]s. 

Former-commit-id: ce88df074bd9e1bae873983971a13b1432dc2a70
Former-commit-id: 7d78b345189222456094eda77478adf6ed582ab7
Former-commit-id: c90aac6b1bc172d5da3851de66f0c8355884e62b
Former-commit-id: 194deb303ddd0a71c1f10c436388a27ef2f0a89a [formerly c145865f89621219cbffd700382dc71f050f0472]
Former-commit-id: 25720b1455eba6d204aecc9fde0cbeb9bbf56d5c