Turn wmi_container into a container.
This defines some types and functions that come from the STL
definition of a container. It's not complete, though, as I do
not see a reason to define size(), max_size(), or swap() at the
moment. (They would be easy to add, if the need ever arises.)

The new #include file might be useful elsewhere as well. I'll look
into that.


Former-commit-id: 1f35c2cce81d6eba3bc63c0ef5edb646a92914b7
Former-commit-id: 38e48bc2ca1f4fd32781017178ee10df4d2d6106
Former-commit-id: 81cc0293c0c83b0bd5d167920eb3f835d630d530
Former-commit-id: 89fc59cb6a6de70b6dfcb1f155f5511dc0964ff5 [formerly 218374f9603997abd9ca34a30ad721f393b009db]
Former-commit-id: 5cf15110affe64e38f7e7cd12c45d3d1938ab67f