filesystem: Ensure SHGetSpecialFolderPathA call uses WINAPI call conv
gfgtdf reported on IRC that building Wesnoth on MSVC++ 2010 and
using a relative path to My Documents with the --config-dir command
line swich results in a runtime error. Paraphrased:

> The value of ESP was not properly saved across a function call

This is caused by mismatched function call conventions between
GetProcAddress()'s return value (__stdcall, a.k.a. WINAPI), and the
function pointer used to store and later call that address (__cdecl).

This commit should fix the issue by ensuring the function pointer is
set to use the WINAPI call convention.


Former-commit-id: d1fe325e8179ec0dd31461b67eb6c47ce26ee471
Former-commit-id: e845e7c9fc93a23769793d22dfbbc182296f28b3
Former-commit-id: d081215edf0192d853d31a8a56e323f5846a5532
Former-commit-id: e3e124b0cc2c99f826ad8d930f1dafe0edcd4e15 [formerly 18bc5d6610346ef429118992bd48a1eca14adde3]
Former-commit-id: 3483eeec9318cea37413fb1d2b8caf9586701626