Maintain difficulty in MP, as it is maintained in SP.
This solves the problem of always seeing a "Medium" difficulty in saves
preview pane for MP campaigns.

N.B. it should not be required to set difficulty in multiple places for
things to work correctly. In the future, it should be made so that
setting difficulty in "game_state::classification().difficulty" would be
enough.


Former-commit-id: d20473dcd89c53fa21329c73664f45c5671b4784
Former-commit-id: 300e0576bded6d0cf7552338ff386569532935d5
Former-commit-id: 5fc90b6221308b137d611a81b3539ee51f96d750
Former-commit-id: dbabe1fd9321260dd6aedda9282cca06dfbdca32 [formerly 15e820cc4ceaa9f6174f94c4cf92bc00d2fc2131]
Former-commit-id: 8b8ac76e8d3b7b81dbbe2e72311deba9e0d280c3