Prevent crash of replay started from start-of-scenario save
Start-of-scenario saves in 1.11 do not contain the [replay_start] tag
any more because its contents are identical to the starting situation.
As a result, the tag needs to be created during the start-up sequence
of a scenario.  Whether this is necessary was previously decided by
checking whether the scenario snapshot was empty.  This works when the
scenario is started by finishing the previous scenario.  However, when
the start-of-scenario savegame is loaded, the snapshot is not empty.
Thus, we need to check whether the snapshot is empty OR whether the
turn number in it is less than 1.


Former-commit-id: aeb136b32a10d7ca3bf8c5f90aaf812b72456f9b
Former-commit-id: e16c796f6ac0ca7e6b798881af70b1496be562d0
Former-commit-id: 849160a55ff38900d05dc0592045e738451d4fd3
Former-commit-id: 90dc5cd4ddc4d59de2558948d9d6cf1e3c1f57fc [formerly 3ea919c508fcb0560e8faf4591b10f69ad5c9f59]
Former-commit-id: bbe9db426e8707f962e903dd49c81e0e47f3daf0