Consider "Hold Position" units in endturn handling.
Do not show the dialogs for missed units if they are told to hold
position.
Changed the string in the dialog that asks about units which have not
been moved at all to be different to the one in the dialog that asks
about units which have still movement left.

Former-commit-id: e12cc2fee661693052e3983e88481e4a9342aa92
Former-commit-id: 1be45f45a36bc5790726ffdd2e797b3ffe2c7d56
Former-commit-id: 16723b11d985488642f7b651c573bfba771918c4
Former-commit-id: e42c8c938399cf2e6ac51606633b7557540c01c1 [formerly aa9458c0dade7abfaca0cfa5d3fa6bce3b86b9cd]
Former-commit-id: df00f702f1e88a419a2ad63019946fba1031f32d