Parameter consistency between can_show() and fire_event().
(These are members of wml_menu_item.)
I opted to get rid of a parameter and use a global. Using globals
is kind of yucky, but in this case, it's basically a choice between
using the global within the procedure or using the global as a
parameter to the procedure. I chose to go with the older function's
approach.


Former-commit-id: 318dae622d7e1374f435ead792743d835ea4fb34
Former-commit-id: fd96a59b7ef01157a962f36048a47b60e5d35ac9
Former-commit-id: 5f538ddf461d4fccc4b92bcc381f00702f2b2756
Former-commit-id: cc3554d55f972fbbe5ebd10dc16fd53d5bb70ef1 [formerly a12bdc0f2f230547e93ffb7e833991d8c9958931]
Former-commit-id: b0892c3e26fed8a3b069abcbf363c5022ec7cf9a