Restore support for negative healing.
The 1.10 implementation did not rule this out, so I've (finally)
tried to restore what 1.10 allowed, with slightly better accounting
of multiple sources of negative healing. I don't know if this is
the best thing to do, but some UMC authors wanted it back. (They
did not want it enough to answer my questions about it, though.)


Former-commit-id: fa801248fb9f174830339862b260f5d37f914da5
Former-commit-id: 4b154915043d7a57dd80c5c47bb9eb9e78d4b947
Former-commit-id: 019864708e4749448021607b4f2ee673e3727a9b
Former-commit-id: 1f4b1790b6e7aaf30a5b48f9848a2e4d624a337e [formerly dd6d413f706e2224f00d3c325cb8d11c7890b7e1]
Former-commit-id: c44570fd60848d80afb284090f38ecd071dadbff