Ignore disabled handlers when checking for existing ids and logging.
Before this update, you could have a one-time-only event with an id
fire, then attempt to add another event with that same id. Whether
or not the second event would stick depended on whether or not the
game was saved and reloaded between the first event firing and the
second event being added.


Former-commit-id: 57d88a432deb232bb7b67e084beaf14b58444d9c
Former-commit-id: ec78a454a58a02abb930c1228ccfac40f6636b55
Former-commit-id: 144e1afc051ac366f8000880e41389062bd54500
Former-commit-id: 47b086abd4fbd8524aedeafd4572160e0744117a [formerly 4364ff15775d6fb7df8d924506fe5529c013d7c3]
Former-commit-id: f0de101a7b62f5825fe4e6415eeaf64fb7ee0688