addon/mg: Check _info.cfg file existence before trying to open it
This is necessary to prevent spurious errors when preparing the Remove
Add-ons dialog such as:

> 20131110 02:21:21 error filesystem: Could not open '/home/shadowm/.wesnoth-1.11/data/add-ons/After_the_Storm/_info.cfg' for reading.

The installation status file is only used to read the original UI name
for the add-on from the add-ons server without establishing a network
connection. If the file doesn't exist we just carry on with an
automatic title generated via string manipulation, so the logged error
or lack thereof when the file does not exist is completely
inconsequential for our ends.


Former-commit-id: da335002fc2010aed1fbcf5de54eb99b1c5e8cf8
Former-commit-id: f0267775e8ab3ad17b6a237ae8be473ee5fe9caf
Former-commit-id: 116c51bb2539ef52b4020099959ad922c819902f
Former-commit-id: 2f48ae0ea4e24e44076fc8039f85efec8fba6673 [formerly 2b22578f7caaaf003de724cd6a2efb8c4f5de929]
Former-commit-id: 9950e77fb552039e0f9c393754090c809c6bfda7