With shared pointers, we no longer need a deep copy.
Technically, this is a behavior change, since copies and assignments
will now share data, causing changes in one copy to affect the other.
However, the logic is that only one wmi_container is active at a time;
the copies and assignments are used to transfer data between scenarios
and to/from saves. So shallow copies are permissable (and slightly
reduce memory consumption).


Former-commit-id: ff89adc7155935a34b476d3e6e5015e8ba44f442
Former-commit-id: bb074cb21e31c7130dd874e2956a513039f36189
Former-commit-id: f5cd3cd47ff34449c10976c62205e0881fb349e3
Former-commit-id: 2836fc5d7c68d77cd6766597fbb86f64c28c56e2 [formerly d92632b383437ffd49c26818c99fcfb3af5ae268]
Former-commit-id: 020b65a7dcfaab7b75c1a83890efa1c70ba87692