When modifying a WML menu item, make a copy first.
This takes advantage of the implicit buffering provided by smart pointers.
If the menu item object is being used, that object will not be changed,
but instead will be destroyed when it is done. Eventually, this will
allow the explicit buffering of menu items to be eliminated, but there
is some functionality to shift around first.


Former-commit-id: bc2307d99ff0e3b8d32a63dc09e533e9af09b62a
Former-commit-id: ee63290f5674840d0cc428488b215a1b8d02946c
Former-commit-id: 738b8619e9e7aea47369ef8fcb9d7fd9b8eae873
Former-commit-id: d5e90ec492888bdd75a3c3bf060319a64cfa22bc [formerly 4517bdb5c380690216dc6c7461b8fe789fa0ddf8]
Former-commit-id: 090d2337a978aaa9dea92233b646f18f0d3d6e1b