Set WML variable turn_number correctly in [pre]start events
There is a C++ function turn() which has value 1 during prestart and
start events.  The same is the case for wesnoth.current.turn in Lua.
By contrast, WML variable turn_number was not updated until after the
start event, resulting in it either being undeclared or containing
whatever the previous value was (such as the last turn number of the
previous scenario).  Now it behaves the same as turn() and
wesnoth.current.turn.


Former-commit-id: dbf6ae1e583df2e28fc23d8b89fd4ae91fef85e9
Former-commit-id: 7cc9674933715ac435e7063f51a64c718679f156
Former-commit-id: 8c8260eafdd3766bb261b8d7055cb0edd81d3ab5
Former-commit-id: de2529a16f5c3c127cf6a2db20efc4454f07d788 [formerly 6d45ed3d987f97d6cfd7e52f89b0b496ab77f0de]
Former-commit-id: 74537e1eb00616159edb5d40e4b3530d192e84fa