SotBE S16: eliminate dependence on luck at scenario start
The difficulty of the scenario depended too much on the success of a
small number of attacks, on whether the player succeeded in killing
one, or even both, of the close enemy leaders on the first turn.
Now, Kapou'e has his own keep, out of reach of the two close enemy
keeps.  Minor changes going along with this are making the two AI
leaders warriors (instead of grunts) and not allowing wolves as
replacements for Jetto and Inarix.


Former-commit-id: 4d3aceff7ab854be01b3b269d18ae98f9072e20f
Former-commit-id: 661b3345121149f5972f25e871be02e7de31e8f0
Former-commit-id: 5cd7bfaaf6a1a4546f6379ffc1d97f21c72f3cce
Former-commit-id: aa02f3cc4bcecf1a7df7744de36414fe19813c16 [formerly 52190a6cde1901ed86e2bc68ca0b02666fdf355a]
Former-commit-id: 5f3674c88470511c602a38524db665f03f7bafc1