scons: Fix test suite linking step after the hotkey refactoring changes
9a41fed5e5edc6dc90e61a9a92712fcfb811cfa1 [formerly 021deb0b659bd2dfe35f58bde06c449262ea8149] [formerly 9e6674034a7559686775364e3237c0b2af86e1f1 [formerly 30741ae1cf9184ee105dbfb5f336179cca358917]] [formerly 66f8047ad510f5ad3150bfc83fe06d9b1115b879] [formerly 8b39eb143a720430ffd00c11b331cd3bfb05cfb4] [formerly 01f6fe443d3046077256073412cac04ea63386f3] adds a few files in replacement
of src/hotkeys.cpp and this somehow broke the test suite target's
linking step resulting in the following errors:

  game_config_manager.cpp:(.text._ZN10test_utils19game_config_managerC2Ev[_ZN10test_utils19game_config_managerC5Ev]+0x4b): undefined reference to `hotkey::manager::manager()'
  game_config_manager.cpp:(.text._ZN10test_utils19game_config_managerC2Ev[_ZN10test_utils19game_config_managerC5Ev]+0x631): undefined reference to `hotkey::manager::~manager()'
  collect2: error: ld returned 1 exit status

This is apparently a linking order issue, since those symbols actually
exist in the object set linked into the test suite executable.

After some fiddling around I discovered that changing the test suite
linking order as follows appears to fix the issue:

Before: libwesnoth_extras.a lua/liblua.a libwesnoth_core.a libwesnoth_sdl.a libwesnoth.a libwesnoth_extras.a lua/liblua.a libtest_utils.a

Now: libtest_utils.a libwesnoth_extras.a lua/liblua.a libwesnoth_core.a libwesnoth_sdl.a libwesnoth.a

The most obvious change is that libtest_utils.a is now linked first.
Additionally, libwesnoth_extras.a and liblua.a are only linked once at
the start. The duplicated entries were introduced in commit
034fd757fd6658964dbc4a79d2e04853721b4cf8 [formerly a82c2a85f76ede979745bef2736c78c9da49667d] [formerly 89e36f62802e2f510c184439f18775728f9003db] [formerly 194c603f6a12b25abeb493381f20aef814a026c7] [formerly 92addd958e7c3e854e153feba80b0482466196ce], but I couldn't really find an
issue with deduplicating them (tested with clang 3.2 and gcc 4.8).

I am not really sure what's going on, but loonycyborg told me to commit
the fix anyway. If Wesnoth spontaneously ignites, you know I am not the
_only_ one to blame.

Nothing of this should affect compilation and linking of the other
targets.


Former-commit-id: 8ac9f13f6bdb82451ab019506db041ccdd9e4e8f
Former-commit-id: 28ae095989d91575a6ba2da94168ad8c4c7ad6f1
Former-commit-id: e2b3bc4ea091a76df411b3ffba315421ec6e3571
Former-commit-id: 6824831c63a5e22e499241bcf71abd08be2347bf [formerly 93f76afd4a176edf1d4571a30ac55ecbe0ec4f08]
Former-commit-id: dd0e43d831cd34469170d43ef7fef3a8cbd39ac0