New lua metatable "wesnoth.mp_settings". Follows pattern of previous lua hooks in src/scripting/lua.cpp.
I would like to be able to access all the fields that occur in
"multiplayer" tag in save game files.

Referring to gamestatus.cpp line 1208, which is what writes
the game state to a config for output, I check if

(game_state.classification().campaign_type=="multiplayer" ),

and if so I ask the mp_settings object for fields.

Referring to line 195 of mp_game_settings.cpp, I use the same names
as appear in save files for the lua hooks.

I don't currently handle vector objects like "active_mods".


Former-commit-id: 1bc98634ade5017b73ba0af74532ba97138714e9
Former-commit-id: 62e5f82e3b320aeba82a54e7a3d31f8714a23c4f
Former-commit-id: f4f9e461f3f15d1b086e010b673d8fa77daa82f7
Former-commit-id: c6179a98ff1395af8167f0f1272ec1ad0b71e622 [formerly 7d45d8c4970f9668237de9fb70cf02ef1d9c4157]
Former-commit-id: 3b44e5b6c16a71a8c6e90f867004f08043f20fa8