create a magic comment for appending custom usage classes
This will keep non-core usage types from being described as a "non-standard usage class" in recruitment error messages. These error messages will only be triggered if there is a mismatch between recruits and recruitment patterns, so this magic comment is low-priority.

It was suggested that wmllint could auto-detect new usage values in unit files, and automatically append them to the list of recognized usage types. This was rejected because of the possibility of adding misspelled or mistaken usage types.


Former-commit-id: 142b36686eba46922648927100b7e5948095a300
Former-commit-id: 707f927e01c0be92717844d19c860bd1b3602ea6
Former-commit-id: 5d491f8944b4323bf6d8229bdfaa7fbe310e309a
Former-commit-id: d468eafbcad34159011bda353ebc38507185df66 [formerly d5e655b80f6b1dcd64a3acdda5524694c480061c]
Former-commit-id: a6b2e22efb1291958f3638415427b621470fa2d5