auto-recognize RECALL characters
Although this is not a core macro (yet - I've been meaning to ask why Wesnoth doesn't have core recall macros), it is used in many UMC projects in a consistent format, making it a de-facto standard.  Test that the macro has only one field, to make sure it isn't deviating from the format.


Former-commit-id: 75617c58942038adea15098cf02454235e04440e
Former-commit-id: e80dc51c1e66a97aa5bc79f5359ad034d3d67656
Former-commit-id: f470e58967b3292a8e42e8b79616577999b1f2c2
Former-commit-id: ecbc3d207b0bb388f1ea6d874e6d71bfba9ef186 [formerly d44bf595204d101bf2494d6eddfe7fb82052782b]
Former-commit-id: b3aebd5e2dcebc3db62720ead8b85206157894e3