Added new CampaignWML attribute "allow_era_choice".
If attribute is set to "no" and campaign with it is selected, eras menu will be
hidden and text to indicate this will be displayed.

Attribute defaults to "yes" and has been added to LoW.

N.B. since current MP code is tighly coupled with eras, the default era will
still be used by the engine to properly initialize stuff. Later on, code should be
improved by making it more modular and independant of eras.


Former-commit-id: 5c78436bf7c444332cac4c557a9a1e5b680f67e3
Former-commit-id: 690d5eb59fbd605433fab15fc7f39546269cd2b1
Former-commit-id: 94e625abf70e2869c7dd0f70733bd48a645c6e2d
Former-commit-id: 79c0d9a265ea62827eafc01ef2f4d70996477423 [formerly 87088605d22b72eaddda2a091f9c7b680b5a8a90]
Former-commit-id: abce8b06192c76defa31096b84bc463c0efbdab1