add RECALL_XY macro to core
Many UMC projects have a macro for this function, but the name isn't
standardized (RECALLXY and RECALL_POS are also in use). However, regardless
of name, all these macros follow the "id x y" order, so for compatibility
I have kept that order. UMC authors started with a RECALL macro and added
coordinates, while the unit-generating macros have to organize a lot of
extra fields that are irrelevant to [recall].


Former-commit-id: 00cc9d49fcab8ed23a1e8ef39b9847719fc20450
Former-commit-id: e79f7275289cab7607e9d897a8a62b4305994f7c
Former-commit-id: d159bee12945770b3e93ad82e01f85b5887bed52
Former-commit-id: 1a9ff1d8e24b43b0133f7cf7a1fee2e6a08760fb [formerly ac0aa42ea9ec39e125d779730e5ca0254d306b88]
Former-commit-id: 108ad19f12cbc23d3b3d5d2dd901d8f2a055cbf7