NR 05a - The Pursuit: add dialogue when a unit steps on the dungeon signpost
Most of the other signposts elicit a remark when a unit steps on them, but
curiously not this one.

Not only do I give the discovering unit a line, I have Tallin speak as
well. This may give alert first-time NR players a clue that this corridor
is important.

Some may prefer that players continue stumbling about, and object.
Unfortunately, there's no NR maintainer or active lord of prose to resolve
any difference of opinion.


Former-commit-id: 926a86f662c0eff54c1945fe03f03c6339681cc3
Former-commit-id: 46cc92ff38dbfe50cbca5e84b80f9b89220c1dba
Former-commit-id: 40d5a2a9e7ced52b8f7c1a2062438619a1b1d0ec
Former-commit-id: 1891fa5d29eabf1323fb2a2672d13ba2f322242f [formerly 09f874f620af56be19e3445638963c719a96e2d2]
Former-commit-id: 78bcef1367f99179e71d0c3d0b2174e3e0fc174e