preferences: Implement a "combo" advanced preferences entry type
Usage example:

    [advanced_preference]
        field=test_combo
        name= _ "Combo test"
        type=combo
        default=option1
        [option]
            id=option1
            name= _ "Option #1"
        [/option]
        [option]
            id=option2
            name= _ "Option #2"
        [/option]
        [option]
            id=option3
            name= _ "Option #3"
        [/option]
    [/advanced_preference]

NOTE:

Now that there are more than two types of advanced preferences entries,
a problem with the saving and restoring of the previous screen surface
contents with the three GUI1 widgets involved (the checkbox, the combo,
and the slider) has appeared, resulting in the advanced preferences
slider glitching into view under certain conditions. This is apparently
related to the widget hiding/unhiding order; for example,
hiding/unhiding the advanced preferences combobox after the slider and
its label results in the combobox glitching into view when switching the
advanced preferences list selection instead.


Former-commit-id: 44f9e32b89b07376165b1898de3b56a74e186fd3
Former-commit-id: 786a78a04d4c800bcccf262d5869ba1dac770898
Former-commit-id: b2f8bfa5d74a00a3b2016af2b4c82288694ba7ed
Former-commit-id: 865a2e39dab547459b82662250cb90e04bdc5b8f [formerly d2478af41c26f6ac6fe2db6cf442b3533d5af095]
Former-commit-id: d55417701a099155d36511500cbd8bdc78e8cceb