preferences: Add description and name_short attributes to combo options
name_short is used for the value labels entered into the main advanced
preferences listbox, which ideally shouldn't be stretched too much
sideways because GUI1 sucks. (Optional, used instead of name.)

description is entered as a second column value in the combo box items
list when available, and should be used as a short description of the
entry. (Optional, used in addition to name in the combo box popup view
only.)


Former-commit-id: cc0ce167c0bb5e6ea7ebfea0be1632d1616c3dbb
Former-commit-id: b80a45ebcf4df432231ec66fd9ccc930eee8c5bd
Former-commit-id: 448f3192d9acc7f1cd696113e2d8c43ad807c3f1
Former-commit-id: 872bf00dcf0cf9b85d9ea9c14861d3b51b8c5759 [formerly 320d9d65f92763d2e8aa1b60ae698d12a07f8ff3]
Former-commit-id: 9d957df099fa4ddc98fff78565a8c1f99fc8b0aa