disable undoing unit creation.
Undoing unit recruits currently ends in oos in most cases, because the random generator is called during unit creation for creating random names.
This is not impossible to fix, ofc we could just enable it for nameless units, but i'm not 100% sure wether units with no names call the random generator for random names or not. Also names normaly don't have any influence on the game (same for genders maybe?), so there is no reason not to be able to undo it from a players point of view. We also cannot just store the random state in undo_stack->add_recruit, because the unit creation happens before that

Former-commit-id: ffe2d21d40886ca8c0e30c6bce28a4a867cfc97b
Former-commit-id: 5bff30a49a0fed7c11e053eef853c02505955eb4
Former-commit-id: 470754db6d70da2d4863ca2786840cda6f65e781
Former-commit-id: 083eacf8544a69b0c9481469fd38f916ca71dae7 [formerly 80a3e6c1ba846b3d2117327fb834d15d95da08c0]
Former-commit-id: 872e3998ecc73403377e79bb4fd0c5925cfc6c3f