wmllint: clean up detection of "illegal children of [if]"
Previously, only lines that parse_attribute could process got passed to
local_sanity_check. So tags survived as ancestors of attribute lines, but
the lines with the tags themselves got discarded. Hence, the "illegal child
of [if]" error was actually triggered by attributes, and the line number
referred to that line rather than the child tag.

While working on [aspect] and [facet] (which originally preceded this commit),
I changed the TypeError exception to pass all lines to local_sanity_check. This
allows for a direct test of the child tags. It turns out many child tags were
not tested because all the attributes were nested deeper inside other tags,
so that the child tag was not the parent of any attribute.

Two core macros failed the new test with the tags [filter_second] and
[filter_second_attack]. This pointed up the fact that these were not in the
list of valid child tags.

It also turns out that tags inside #ifdefs didn't fit the original test,
because "#ifdef" was considered an ancestor, interrupting the link between
"[if]" and the child tag.


Former-commit-id: 78da34f6a2f58ad596cb012c429a91ac9a6ed29b
Former-commit-id: 49b913a51b81f18f04e5a085b7257f08d191c498
Former-commit-id: 75882e35d3f014ef9e8f75420fb4b556d6259f7d
Former-commit-id: 0723e1042d918c7a6d7ef3bd5f741077cb5e2437 [formerly 49d6ca6fbfa92a7b29a1817e7534066ad2588ce2]
Former-commit-id: a9dce95ab53da6859f982741ea86f300f39b0fb2