preferences: Prevent ghost widgets glitch in Advanced by hiding all first
Hiding all Advanced preferences control widgets before deciding which to
unhide for the current selection avoids the situation where a previously
hidden widget is unhid and captures the background including a visible
widget about to be unhidden, resulting in the newly visible widget
restoring the previous widget's surface when drawn.

The bug this commit fixes only became possible when we gained a third
advanced preferences control widget/advanced preferences option type. It
is possible that the widget construction/placement/visibility toggling
orders are mismatched, but to be honest I'm not too interested in
messing with that further given this sensible workaround.


Former-commit-id: a49b9298acc9184a98634b7eea1c46247596bfce
Former-commit-id: 4022a058f156a6a9d0032d56b326301e5610c4f7
Former-commit-id: 4504151be7034a48549e434b0e816e0e7215f685
Former-commit-id: 241392ef3f33342c2cf854b77306175e36872ada [formerly 8f6d2119bba09919ec8a35fd914325e29c28143b]
Former-commit-id: 2bc85df40da16c27d3b4cb51d15636e34fa9856b