campaignd: Abort sending if the pack size is negative (bug #20893)
Sending an add-on whose size could not be determined means that the pack
in which it is stored is inaccessible in some way. Proceeding in this
situation, with the network::send_file() fix for the same condition,
results in the built-in add-ons client stalling forever waiting for
campaignd to send data to it, so we must respond with an [error] tag
instead.

This is the final step in fixing bug #20893.


Former-commit-id: 0600f8f99b10ee4cd2cf94ca425b91d3682bce66
Former-commit-id: ab207525c86abab3d812f9c91d4719e9ddc51011
Former-commit-id: 07670451b16122f12fa2be2cad67646b2a34bd69
Former-commit-id: 583f3ff943ac620058d84f15ed57d3df672ff377 [formerly 99bf7a9987ce46f41433aff2c7b209e347e5f76e]
Former-commit-id: 62cc8607481d60b006c631fd20f72988ec962bd6