campaignd: Abort sending if the pack size is negative (bug #20893)
Sending an add-on whose size could not be determined means that the pack
in which it is stored is inaccessible in some way. Proceeding in this
situation, with the network::send_file() fix for the same condition,
results in the built-in add-ons client stalling forever waiting for
campaignd to send data to it, so we must respond with an [error] tag
instead.

This is the final step in fixing bug #20893.


Former-commit-id: 3c6f9ab47f4f601d8e59ecb050ac6b17295cd926
Former-commit-id: c8d385c1451f0dd66a0ecba1b9cb99f4cf9b8068
Former-commit-id: 3d316896c38110c8817df3a2ebcfd62e69ccd3fc
Former-commit-id: 3705dc8cf9b04101d87045d2d65b51bc959427a4 [formerly ea08cca267b536dd8166cdc8ad0afee7c04685db]
Former-commit-id: 720c0eaa42244b9297259c15391eec2e78d4d98b