wmltools.py: fix macro call parsing to prevent wmlscope stringliteral mismatch
I am responsible for wmlscope suddenly giving macro mismatch errors, as was
pointed out on IRC. The tell is that in each of these errors, the last field
is interpreted as a stringliteral.

An argument that starts and ends with a quote is interpreted as a stringliteral.
But under previous code, that couldn't happen: quote characters toggled the
instring status on and off, but weren't added to arg, so no arg contained any
quotes. I didn't like the fact that this stripped the quotes out of translatable
strings, so I changed the code to include quotes in arg and strip them out if
they began and ended the arg. Unfortunately, while I stripped out quotes when
arg was terminated by a space, I didn't think it was necessary when the right
bracket terminated the arg. It was.


Former-commit-id: 8d9fe7f021a9870441d89c60f975800e9c24271b
Former-commit-id: 235b710d98f2111870c9142543f7f1ec4484a661
Former-commit-id: baf2b1297ce800aa92368a39c21eb8cb945bbea6
Former-commit-id: 1e9687864ace49f306b95e6057833312d37ea911 [formerly f96e458e42975c9258ebf34308b68a8f44f3eaa9]
Former-commit-id: 6f7b92715f2d1170e5499613e387285af51816cb