UtBS: add and remove "wmllint: who" and "recognize" magic comments
The main character macros now have "who is" pairings. This, along with the
advent of auto-recognition of  NAMED_*UNIT ids, makes some "recognize" magic
comments unnecessary. Meanwhile, the new "unknown speaker" warning makes
adding others necessary.

The ally characters still trigger "unknown speaker" warnings. There are two
possible solutions: one, to add a set of four "recognize" mcs to several
scenarios. The second, which I lean towards, would be to create a macro
for the recall of the ally character, and pair it with a "who" mc. I will
leave the choice to the UtBS maintainer, however.


Former-commit-id: d201292ecad94dc912c4fbd3017aa193809b2165
Former-commit-id: 97c56314a6ae9b0f3e19c6e9a09a84b43b8ad850
Former-commit-id: 60632aa8d69ea599f58d825ad9e1c0c87b0661ed
Former-commit-id: b3f658900f2074f60545a99d0bd15badec755f6e [formerly 5a7e79e749c3dfbcdd359f5cf02129d4423842c7]
Former-commit-id: 1046e2f47ac29774c0bb6998c4a0548cd43d562a