HttT: remove obsolete "wmllint: recognize" magic comments
The exception is Gryphon Mountain, where the magic comments are updated to
recognize the awakened ids.

This is not the only approach that could be used. wmllint could be upgraded
to recognize the transformed ids in [modify_unit]. This is more involved
than I want to do right now, for what I expect to be a corner case.

The other way would be to also filter the awakening events by type as well
as id, and not change the id at all. The original awakening code killed the
old unit and created a new one in its place, and the new id distinguished
the new unit from the old one. However, now that we are modifying an existing
unit, we do not need to change ids, especially since this is not seen by
the player.

For now, though, the easiest path is just to tweak the already existing
magic comments.


Former-commit-id: da1aed8cf3fc4800a9a7c32d59ea2c84e4f5a0bf
Former-commit-id: 15db2f92c3131748aeda77f26253624480f8263c
Former-commit-id: 603252bbb3d0092870be9f21699b4452f294274f
Former-commit-id: 28da45a6825c1cd7db302e62a74f21d2e53060dc [formerly 4caa7e5ec03e29b4f21618087a6258e89ad69434]
Former-commit-id: 9dc1814c76bf9344797a19cc34aea45d0a3e4b2c