ai.cfg: fix "#ifdef MEDIUM" to NORMAL in attack_depth
Though this may be in vain, as I read on the wiki that attack_depth is not
working with the RCA ai and is deprecated. While this bug might break
attack_depth on one difficulty level, it shouldn't have broken it entirely,
so this is probably not the reason it doesn't seem to be working.


Former-commit-id: 8e48f18176a45835408dd7b870c3ebb10b336e8e
Former-commit-id: 0d84c6b7c0e96b5d1813712b5c81aac5c0209fcd
Former-commit-id: 44b8340b7016768f964fa30a3a52a3407676f919
Former-commit-id: 1fe563ffbc99571c928536739cf16c5d92e07dd5 [formerly fed602c96eae41b8599053d09a9bf12d0cb7dd76]
Former-commit-id: 04ef50bc1fdf5f2e784ce9be39024d365dded9b1