ai.cfg: fix "#ifdef MEDIUM" to NORMAL in attack_depth
Though this may be in vain, as I read on the wiki that attack_depth is not
working with the RCA ai and is deprecated. While this bug might break
attack_depth on one difficulty level, it shouldn't have broken it entirely,
so this is probably not the reason it doesn't seem to be working.


Former-commit-id: 2b88af23fab6bf70511f4fe88540ce9e8f919d42
Former-commit-id: 8295eded437830198a368e8b8f08eafca23f1a10
Former-commit-id: 7f43965143d9dc452c7e55f09625b511b91af5da
Former-commit-id: e0abbfdfc75407f47f3c86de9a80e6a109321860 [formerly c32ca43d993926db994c69f3604d16e1d665176e]
Former-commit-id: 591a87d6b85c4f3b68f587377bcf7f279a249d82