about: Split multi-line section headers for the credits screen
Multi-line section headers such as those generated by campaigns with
multi-line titles result in slight scrolling glitches because the
scrolling code assumes every entry in the text vector to comprise a
single line.

We don't want to split those headers for the help screen because the
current approach results in nicely formatted headers for the
Contributors page there, so a hacky compromise for now should be to only
split multiline headers for the scrolling credits.

An alternative would have been to make the text vector a linked list so
about::show_about could deal with this issue directly, but that seems
rather wasteful.

Former-commit-id: 97e206ee1269bd7c73f051843fcbc17690e5180b
Former-commit-id: 7d917f5eeeda92ce1ec6568ae20ec549f55f3e68
Former-commit-id: be40f825645c59db83d357c25900b94d4aa98aae
Former-commit-id: a2c3cf52dff72456d9172134bb5258a366043f9b [formerly de144a7a3743935732e50fc0b3a05846b4ef6e57]
Former-commit-id: eecd126a58618339113a5f6c8925cb94fbe98ce7