display: Drop trailing underscore from the color_adjust_blue_ WML option
This was used in saved games during the serialization of the display
configuration (see the display::write() method) since around 1.10.x. The
trailing underscore was clearly not intentional, as the siblings of this
attribute for the red and green channels don't have it, so drop it.

Resulting from this commit, saved games from previous versions that did
have the trailing underscore and were affected by a [color_adjust]
action may be loaded with the wrong color channel configuration (blue
will be reset to the default of zero), but this situation should be rare
enough to not absolutely require backwards-compatibility code.


Former-commit-id: 151c503ac486ab80a3c3d3461ab748301d04ffea
Former-commit-id: dc42932ba3f960d10ab740c90f360bb66c4febf9
Former-commit-id: 3d91a771036ddfe0c783bb2b765a4e1e829a35f6
Former-commit-id: 15b9ca42c2715c23001ad9937f3ff9715ff58f9a [formerly cc9cca9e870f016e006518b72b1207d25bb717ea]
Former-commit-id: fce49e31a7850d48b84cfb9032c16df375fdd172