sound: Call Mix_HaltChannel() directly for muted channels (bug #21426)
As described in detail in the bug report <https://gna.org/bugs/?21426>,
SDL_mixer's notion of a channel that's "not playing" appears to include
channels with their volume set to zero (a direct consequence of setting
Wesnoth's sound effects volume to zero in preferences or through the
[volume] WML action), which results in Mix_FadeOutChannel() doing
nothing, letting the channel resume playing at a later time once the
volume is reset.

For sound sources removed with [remove_sound_source] while the sound
effects volume is zero, this effectively means that once the volume is
reset, the sound sources will continue playing in SDL_mixer-land and
Wesnoth (or more specifically, WML) can't do anything about it because
the associated data structures have been released and the sound sources
don't exist in Wesnoth-land anymore.

Calling Mix_HaltChannel() directly when the channel's volume is already
set to zero avoids this issue and does not negate the point of the
fade-out effect since there is no way to fade out a channel when it's
muted anyway.


Former-commit-id: 0ee6e15c14daeff052e1e09083b7312c463cace1
Former-commit-id: 109150d8a60c2bdc89c2e21a9bb39b6171a64584
Former-commit-id: 1b4ef8b4a7de653918ef2e4dd35b4ef98cf3f969
Former-commit-id: f98905eaca0cf620c4a6422c6c2ddb009945714e [formerly ec74d0a16cc80f1bcd8b3801d0f742e431af0e4e]
Former-commit-id: 0aa1de2eff9390cb940d2e4df7da78a3e541269a